Remove `AddAutoMapper` from services configuration
As it seems, this line has absolutely no effect on the code, since all
toggle endpoints seem to work as expected, in a first glance.

And yet, it is causing a massive failure in tests, which I haven't been
able to fully understand. The following links provide more details about
the issue:
- https://stackoverflow.com/q/1091853
- https://github.com/Microsoft/vsts-tasks/issues/4514
- https://github.com/Microsoft/vsts-tasks/issues/4684
- https://github.com/Microsoft/vsts-tasks/pull/5622

Then, I simply lost track of it ¯\_(ツ)_/¯