Replace AutoMapper with plain converters
This change removes AutoMapper dependency from the project.

It was identified that the dependency injection setup of AutoMapper was
in fact required. Since its usage was causing a rather tough to handle
problem in the tests, it was decided to drop AutoMapper usage, in favor
of plain converters.

This change adds the converter interface, as well as it updates the
proper service implementations, in order to use this new interface.